pyproject.toml
[project]
name = "cm-agents"
version = "0.1.0"
description = "Sistema de agentes para automatizar diseños de redes sociales"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.60.0",
    "openai-agents>=0.1.0",
    "mcp>=1.7.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "typer[all]>=0.12.0",
    "rich>=13.0.0",
    "pillow>=10.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    # API server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=14.0",
    "python-multipart>=0.0.18",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
cm = "cm_agents.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cm_agents"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "-v --tb=short"
